+++ src/configuration/subgraph/auth/mod.rs
@@ -2,5 +2,10 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct ServiceAuth {
-    identifier: String,
+    pub identifier: String,
+    pub requesting_party: String,
+    pub requesting_party_name: String,
+    pub requesting_party_origin: String,
+    /// The name of the data source to save credentials.
+    pub data_source: String,
 }